refactor(component): FriendSelector, GiftCategorySelector, PromoBanner, RealtimeGiftRank 컴포넌트들을 components/ 디렉토리에서 components/Home 디렉토리로 이동함
