+++ src/components/Home/FriendSelector.tsx
