+++ src/pages/Home.tsx
@@ -1,8 +1,8 @@
 import Layout from '../components/Layout';
 import NavBar from '../components/NavBar';
-import FriendSelector from '../components/FriendSelector';
-import GiftCategorySelector from '../components/GiftCategorySelector';
-import RealtimeGiftRank from '../components/RealtimeGiftRank';
+import FriendSelector from '../components/Home/FriendSelector';
+import GiftCategorySelector from '../components/Home/GiftCategorySelector';
+import RealtimeGiftRank from '../components/Home/RealtimeGiftRank';
 
 function Home() {
   return (
